+++ CHANGELOG.md
@@ -10,7 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### VB -> C#
 
-* CInt wrongly converted to (int)[#658](https://github.com/icsharpcode/CodeConverter/pull/658)
+* Improve conversion accuracy for CInt [#658](https://github.com/icsharpcode/CodeConverter/pull/658)
+* Create non-static delegate when converting a shared event[#671](https://github.com/icsharpcode/CodeConverter/issues/671)
 
 ### C# -> VB
 * Generics type parameter missed BC30737 and BC32042 [#682] (https://github.com/icsharpcode/CodeConverter/issues/682)